Allow compilation in multiple translation units
Tests did not compile before this commit because there are two C++
source files that defined a static class member, violating ODR.
Worked around by using a the Myers singleton trick to store the
pointer, but allows it to be set on the first call.
There must be a better way. Sadly, we need a global somewhere as
you cannot pass the calling object to the ASIO API. I cannot think
of another way to do this in a header-only lib.